backend/pyproject.toml
[project]
name = "passportai-backend"
version = "0.1.0"
description = "PassportAI DPP Operations Module - Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.32.0,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "psycopg2-binary>=2.9,<3.0",
    "alembic>=1.14,<2.0",
    "pydantic[email]>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "python-dotenv>=1.0,<2.0",
    "bcrypt>=4.0,<5.0",
    "PyJWT>=2.8,<3.0",
    "python-multipart>=0.0.9,<1.0",
    "httpx>=0.27,<1.0",
    "pymupdf>=1.25,<2.0",
    "pytesseract>=0.3.10,<1.0",
    "Pillow>=10.0,<12.0",
    "anthropic>=0.42,<1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
